requirements.txt
sockjs-tornado==0.0.4
tornado==2.2
redis
